README: Hoist PGO packages into list of packages to install
Originally, I had listed the PGO packages as an aside because they were
just for building the kernels and I did not want to bloat people's
systems.

However, it has caused some headaches over time because they did not
realize that they needed more packages than the ones listed because the
PGO packages were listed in a paragraph below and it was not very clear.

Fix this by just moving the packages needed for --pgo to work into the
main list of packages; it won't hurt people. Add libelf to the list to
make x86 defconfig work due to the ORC unwinder.

Suggested-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>