build-llvm.py: Improve '--use-good-revision' text
Make it clear that this is an option that needs to be used with caution:

1. It is hard to flip around on branches due to an implied
   '--single-branch' during cloning by default.

2. It is hard to checkout to just a SHA1 (like when using
   '--use-good-revision' or tracking down an issue in the
   middle of master).

This is really an option designed for CIs where a one off clone is
needed for speed purposes. Manually managing the repo and taking the
script's heuristic out of it is usually a better option.

While we're here, alphabetize the position of '--shallow-clone'.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>