build-llvm.py: Support '--branch' with '--shallow-clone'
As it stands now, '--shallow-clone' does not allow switching to a branch
other than master. This is because '--depth=1' implies '--single-branch'
so we only fetch the default (master).

To support switching branches with '--shallow-clone', we add
'--no-single-branch' if the user specifies a branch other than master.

There is one snag though: this does not allow a user to switch branches
if they clone with '--shallow-clone' because '--[no-]single-branch' is a
cloning time decision. There will be a future patch that improves the
error reporting as well as help option to let users know that unless
they are seriously bandwidth restricted, they should be doing a full
clone.

Closes: #81
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>