kernel/build.sh: Add toolchain version information
This is helpful for making sure that the right tools are being used for
the build. This was inspired by our CI, where the Ubuntu image has
clang installed:

https://help.github.com/en/actions/reference/software-installed-on-github-hosted-runners

We want to make sure that we are explicitly using the version that we
built so display this information to the user.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>